Be more lenient in accepted instants
When we use the Joda `DateTime` type, we accept a lot of different
inputs that are not complete instant. Instant without month, day or time
information are accepted and default to the beginning of the missing
period.

This adds a more lenient parser (keeping the strict serializer) and
ports some of the Joda time tests to ensure we have compatibility in
accepted inputs.