temporary fallback to joda time when cannot parse instant
During a certain period of time, when we fail to parse an `Instant` from
the user input, fallback to joda time to make sure we don't break any
user implementation.
We log the error to make sure that we take care of the case that we are
missing.
After a period of time where we don't have any error, we can remove the
fallback.